Modifica installazione gradle e prova di esecuzione main

+++ backend/springboot/src/main/java/com/AssignmentTWEB/springboot/model/Movie.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*; //JPA annotations
 public class Movie {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) //auto-increment ID
-    private Long id_movie;
+    private int id_movie;
 
     @Column(nullable = false) //Not Null
     private String name;
